wip: add client for bifrost, add fake
* create resources with k8sclient when running locally, use bifrost api in prod
